Removed. Experiment to use Google Code svn, via webdav, as repo for OpenWrap failed.
Reading anonymous, or logged in, of svn repo at Google Code works like a charm. Writing however (via logged in webdav) seems to work, but, on Mac OS X with a error and a 0-byte-file as a result, and with netdrive, without any message and without any result.

Thus we cannot use this to publish wraps.